Add custom editor to storybook using new editor system
