+++ docs/storybook/src/frontstage/ToolSettings.tsx
@@ -12,6 +12,8 @@ import {
   StandardLayout,
   WidgetState,
 } from "@itwin/appui-react";
+import { EditorsRegistryProvider } from "@itwin/components-react-internal/src/components-react/new-editors/editors-registry/EditorsRegistryProvider";
+import { TagsEditorSpec } from "src/tools/TagEditorNew";
 
 interface Props
   extends Pick<
@@ -21,47 +23,51 @@ interface Props
   mode?: "widget" | "floating";
 }
 
+const editors = [TagsEditorSpec];
+
 /** [FrontstageProvider](https://www.itwinjs.org/reference/appui-react/frontstage/frontstageprovider/) can be used to configure a frontstage. */
 export function ToolSettingsStory(props: Props) {
   const { mode, ...rest } = props;
   return (
-    <AppUiStory
-      layout="fullscreen"
-      demoIModel={{ default: "blank" }}
-      frontstages={[
-        createFrontstage({
-          contentGroupProps: {
-            id: "ViewportContentGroup",
-            layout: StandardContentLayouts.singleView,
-            contents: [
-              {
-                id: "ViewportContent",
-                classId: IModelViewportControl,
-              },
-            ],
-          },
-          hideToolSettings: false,
-          toolSettings: {
-            id: "toolSettings",
-            defaultState:
-              mode === "floating" ? WidgetState.Floating : undefined,
-          },
-          layout: (
-            <StandardLayout
-              toolSettings={{
-                defaultLocation:
-                  mode === undefined
-                    ? undefined
-                    : {
-                        location: StagePanelLocation.Right,
-                        section: StagePanelSection.Start,
-                      },
-              }}
-            />
-          ),
-        }),
-      ]}
-      {...rest}
-    />
+    <EditorsRegistryProvider editors={editors}>
+      <AppUiStory
+        layout="fullscreen"
+        demoIModel={{ default: "blank" }}
+        frontstages={[
+          createFrontstage({
+            contentGroupProps: {
+              id: "ViewportContentGroup",
+              layout: StandardContentLayouts.singleView,
+              contents: [
+                {
+                  id: "ViewportContent",
+                  classId: IModelViewportControl,
+                },
+              ],
+            },
+            hideToolSettings: false,
+            toolSettings: {
+              id: "toolSettings",
+              defaultState:
+                mode === "floating" ? WidgetState.Floating : undefined,
+            },
+            layout: (
+              <StandardLayout
+                toolSettings={{
+                  defaultLocation:
+                    mode === undefined
+                      ? undefined
+                      : {
+                          location: StagePanelLocation.Right,
+                          section: StagePanelSection.Start,
+                        },
+                }}
+              />
+            ),
+          }),
+        ]}
+        {...rest}
+      />
+    </EditorsRegistryProvider>
   );
 }
