+++ docs/storybook/src/tools/CustomEditorTool.tsx
@@ -2,21 +2,14 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import * as React from "react";
+
 import {
   BasePropertyEditorParams,
   DialogItem,
   DialogPropertySyncItem,
   PropertyDescription,
-  PropertyValueFormat,
 } from "@itwin/appui-abstract";
 import { EventHandled } from "@itwin/core-frontend";
-import { Tag, TagContainer } from "@itwin/itwinui-react";
-import {
-  PropertyEditorBase,
-  PropertyEditorProps,
-  TypeEditor,
-} from "@itwin/components-react";
 import { StoryPrimitiveTool } from "./StoryTool";
 
 const properties = {
@@ -88,14 +81,23 @@ export class CustomEditorTool extends StoryPrimitiveTool {
   }
 }
 
-export class CustomTagsPropertyEditor extends PropertyEditorBase {
-  public get reactNode(): React.ReactNode {
-    return <CustomTagsEditor />;
-  }
+// Custom property editor param for tags.
+export interface TagsParam extends BasePropertyEditorParams {
+  type: "custom-tags-param";
+  tags: TagData[];
 }
 
+// Helper to create custom tags param.
+export function createTagsParam(tags: TagData[]): BasePropertyEditorParams {
+  return {
+    type: createTagsParam.type,
+    tags,
+  } as TagsParam;
+}
+createTagsParam.type = "custom-tags-param";
+
 // Data structure for tags
-interface TagData {
+export interface TagData {
   id: string;
   label: string;
 }
@@ -114,73 +116,3 @@ const tagsStore: TagData[] = [
     label: "Tag 3",
   },
 ];
-
-// Custom property editor param for tags.
-interface TagsParam extends BasePropertyEditorParams {
-  type: "custom-tags-param";
-  tags: TagData[];
-}
-
-// Helper to create custom tags param.
-function createTagsParam(tags: TagData[]): BasePropertyEditorParams {
-  return {
-    type: createTagsParam.type,
-    tags,
-  } as TagsParam;
-}
-createTagsParam.type = "custom-tags-param";
-
-// eslint-disable-next-line react-refresh/only-export-components
-const CustomTagsEditor = React.forwardRef<TypeEditor, PropertyEditorProps>(
-  (props, ref) => {
-    const { propertyRecord } = props;
-    const elRef = React.useRef<HTMLDivElement>(null);
-    React.useImperativeHandle(
-      ref,
-      () => ({
-        getPropertyValue: async () => {
-          return undefined;
-        },
-        htmlElement: null,
-        hasFocus: false,
-      }),
-      []
-    );
-    const tags = React.useMemo(() => {
-      if (!propertyRecord) return [];
-      const params = propertyRecord.property.editor?.params;
-      const tagsParam = params?.find((param): param is TagsParam => {
-        return param.type === createTagsParam.type;
-      });
-      return tagsParam?.tags ?? [];
-    }, [propertyRecord]);
-
-    return (
-      <TagContainer ref={elRef}>
-        {tags.map((tag) => (
-          <Tag
-            key={tag.id}
-            style={{
-              blockSize: "var(--iui-size-l)",
-            }}
-            onRemove={() => {
-              if (!props.propertyRecord) return;
-
-              const tagIds = tags.map((t) => t.id);
-              const newTagIds = tagIds.filter((id) => id !== tag.id);
-              props.onCommit?.({
-                propertyRecord: props.propertyRecord,
-                newValue: {
-                  valueFormat: PropertyValueFormat.Primitive,
-                  value: JSON.stringify(newTagIds),
-                },
-              });
-            }}
-          >
-            {tag.label}
-          </Tag>
-        ))}
-      </TagContainer>
-    );
-  }
-);
